This commit marks the end of efficiency concerns
This is the last part of tonight's coding where joe will give the
slightest fucking concern to efficiency. If I have to get the data I
want by parsing through this sequence in O(n^500), so be it. I am
running through it before I get up from this computer. If it's O(n^500),
then I will die trying.