Song Select Works... So long as you pick the right song.
So Song Select works but it turns out that our midi code is incompatible
with a lot of midi files.

For Like a Prayer, Mountain King, and New World it throws some kind of
error about not being able to find the end of the midi file.

For Symphony No 5 something is out of bounds in the midi file or
something.

The others work fine though! So that's good!